Eri/improvements (#3)
* launchsettings cleanup

* adds DB migrations at startup

* fixes path to projects in sln

* fixes invalid Dapr component name

* fixing DBcontext errors due to invalid data type

* cleanup and defaults

+++ src/SpyrosoftLearn/Controllers/RoundsController.cs
@@ -91,7 +91,7 @@ namespace SpyrosoftLearn.Controllers
         }
 
         // GET: Rounds/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Rounds == null)
             {
@@ -171,7 +171,7 @@ namespace SpyrosoftLearn.Controllers
         // POST: Rounds/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             if (_context.Rounds == null)
             {
@@ -187,7 +187,7 @@ namespace SpyrosoftLearn.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool RoundExists(long id)
+        private bool RoundExists(int id)
         {
           return (_context.Rounds?.Any(e => e.Id == id)).GetValueOrDefault();
         }
